+++ lib/cocoapods/resolver.rb
@@ -85,9 +85,12 @@ module Pod
     #
     def resolve
       dependencies = @podfile_dependency_cache.target_definition_list.flat_map do |target|
-        @podfile_dependency_cache.target_definition_dependencies(target).each do |dep|
-          next unless target.platform
-          @platforms_by_dependency[dep].push(target.platform)
+        @podfile_dependency_cache.target_definition_dependencies(target).tap do |deps|
+          unless target.platform.nil?
+            deps.each do |dep|
+              @platforms_by_dependency[dep].push(target.platform)
+            end
+          end
         end
       end.uniq
       @platforms_by_dependency.each_value(&:uniq!)
@@ -111,6 +114,7 @@ module Pod
           explicit_dependencies = @podfile_dependency_cache.target_definition_dependencies(target).map(&:name).to_set
 
           used_by_aggregate_target_by_spec_name = {}
+          explicitly_included_by_aggregate_by_spec_name = {}
           used_vertices_by_spec_name = {}
 
           # it's safe to make a single pass here since we iterate in topological order,
@@ -122,6 +126,7 @@ module Pod
             if explicitly_included || vertex.incoming_edges.any? { |edge| used_vertices_by_spec_name.key?(edge.origin.name) && edge_is_valid_for_target_platform?(edge, target.platform) }
               validate_platform(vertex.payload, target)
               used_vertices_by_spec_name[spec_name] = vertex
+              explicitly_included_by_aggregate_by_spec_name[spec_name] = vertex if explicitly_included
               used_by_aggregate_target_by_spec_name[spec_name] = vertex.payload.library_specification? &&
                 (explicitly_included || vertex.predecessors.any? { |predecessor| used_by_aggregate_target_by_spec_name.fetch(predecessor.name, false) })
             end
@@ -132,7 +137,9 @@ module Pod
               payload = vertex.payload
               non_library = !used_by_aggregate_target_by_spec_name.fetch(vertex.name)
               spec_source = payload.respond_to?(:spec_source) && payload.spec_source
-              ResolverSpecification.new(payload, non_library, spec_source)
+              transitive = !explicitly_included_by_aggregate_by_spec_name.fetch(vertex.name, false)
+              incoming_spec_vertices = vertex.incoming_edges.map { |edge| edge.origin.payload }
+              ResolverSpecification.new(payload, non_library, spec_source, incoming_spec_vertices, transitive)
             end.
             sort_by(&:name)
         end