+++ lib/cocoapods/resolver/resolver_specification.rb
@@ -17,10 +17,28 @@ module Pod
       attr_reader :used_by_non_library_targets_only
       alias used_by_non_library_targets_only? used_by_non_library_targets_only
 
-      def initialize(spec, used_by_non_library_targets_only, source)
+      # @return [Bool] whether this resolved specification is a transitive dependency that was not directly included
+      #         by the target that depends on it
+      #
+      attr_reader :transitive
+      alias :transitive? transitive
+
+      # @return [Array<Specification>] specifications which depend on this specification
+      #
+      attr_reader :dependent_specs
+
+      # @param [Pod::Specification] @see #spec
+      # @param [Boolean] used_by_non_library_targets_only @see #used_by_non_library_targets_only
+      # @param [Pod::Source] source @see #source
+      # @param [Array<Specification>] dependent_specifications @see #dependent_specifications
+      # @param [Boolean] transitive @see #transitive
+      #
+      def initialize(spec, used_by_non_library_targets_only, source, dependent_specs, transitive)
         @spec = spec
         @used_by_non_library_targets_only = used_by_non_library_targets_only
         @source = source
+        @dependent_specs = dependent_specs
+        @transitive = transitive
       end
 
       def name