+++ lib/cocoapods/target/build_settings.rb
@@ -924,7 +924,7 @@ module Pod
         #-------------------------------------------------------------------------#
       end
 
-      # A subclass that generates build settings for a `PodTarget`
+      # A subclass that generates build settings for an `AggregateTarget`
       class AggregateTargetSettings < BuildSettings
         #-------------------------------------------------------------------------#