+++ spec/unit/installer/analyzer_spec.rb
@@ -1614,8 +1614,8 @@ module Pod
 
       describe '#group_pod_targets_by_target_definition' do
         it 'does include pod target if any spec is not used by tests only and is part of target definition' do
-          spec1 = Resolver::ResolverSpecification.new(stub(:root => stub(:name => 'Pod1')), false, nil)
-          spec2 = Resolver::ResolverSpecification.new(stub(:root => stub(:name => 'Pod1')), true, nil)
+          spec1 = Resolver::ResolverSpecification.new(stub(:root => stub(:name => 'Pod1')), false, nil, [], false)
+          spec2 = Resolver::ResolverSpecification.new(stub(:root => stub(:name => 'Pod1')), true, nil, [], false)
           target_definition = @podfile.target_definitions['SampleProject']
           pod_target = stub(:name => 'Pod1', :target_definitions => [target_definition], :specs => [spec1.spec, spec2.spec], :pod_name => 'Pod1')
           resolver_specs_by_target = { target_definition => [spec1, spec2] }
@@ -1623,16 +1623,16 @@ module Pod
         end
 
         it 'does not include pod target if its used by tests only' do
-          spec1 = Resolver::ResolverSpecification.new(stub(:root => stub(:name => 'Pod1')), true, nil)
-          spec2 = Resolver::ResolverSpecification.new(stub(:root => stub(:name => 'Pod1')), true, nil)
+          spec1 = Resolver::ResolverSpecification.new(stub(:root => stub(:name => 'Pod1')), true, nil, [], false)
+          spec2 = Resolver::ResolverSpecification.new(stub(:root => stub(:name => 'Pod1')), true, nil, [], false)
           target_definition = stub('TargetDefinition')
-          pod_target = stub(:name => 'Pod1', :target_definitions => [target_definition], :specs => [spec1.spec, spec2.spec], :pod_name => 'Pod1')
+          pod_target = stub('PodTarget', :name => 'Pod1', :target_definitions => [target_definition], :specs => [spec1.spec, spec2.spec], :pod_name => 'Pod1')
           resolver_specs_by_target = { target_definition => [spec1, spec2] }
           @analyzer.send(:group_pod_targets_by_target_definition, [pod_target], resolver_specs_by_target).should == { target_definition => [] }
         end
 
         it 'does not include pod target if its not part of the target definition' do
-          spec = Resolver::ResolverSpecification.new(stub(:root => stub(:name => 'Pod1')), false, nil)
+          spec = Resolver::ResolverSpecification.new(stub(:root => stub(:name => 'Pod1')), false, nil, [], false)
           target_definition = stub
           pod_target = stub(:name => 'Pod1', :target_definitions => [], :specs => [spec.spec])
           resolver_specs_by_target = { target_definition => [spec] }
@@ -1686,6 +1686,31 @@ module Pod
           should.raise(Informative) do
             @analyzer.analyze
           end.message.should.include 'The subspecs of `AFNetworking` are linked to different build configurations for the `Pods-SampleProject` target. CocoaPods does not currently support subspecs across different build configurations.'
+        end
+
+        it 'does not implicitly whitelist a whitelisted dependency\'s transitive dependencies' do
+          SpecHelper.create_sample_app_copy_from_fixture('SampleProject')
+          @podfile = Podfile.new do
+            platform :ios, '8.0'
+            project 'SampleProject/SampleProject'
+            target 'SampleProject' do
+              pod 'AFNetworking'
+              pod 'RxCocoa', '4.0.0', :configurations => ['Debug']
+            end
+          end
+          @analyzer = Installer::Analyzer.new(config.sandbox, @podfile)
+          target_definition = @podfile.target_definitions['SampleProject']
+          aggregate_target = @analyzer.analyze.targets.find { |t| t.target_definition == target_definition }
+
+          aggregate_target.pod_targets_for_build_configuration('Debug').map(&:name).should == %w(
+            AFNetworking
+            RxCocoa
+            RxSwift
+          )
+
+          aggregate_target.pod_targets_for_build_configuration('Release').map(&:name).should == %w(
+            AFNetworking
+          )
         end
       end